generateBountyName + make bounty.name non null
