+++ apps/web/app/(ee)/api/bounties/route.ts
@@ -1,4 +1,5 @@
 import { recordAuditLog } from "@/lib/api/audit-logs/record-audit-log";
+import { generateBountyName } from "@/lib/api/bounties/generate-bounty-name";
 import { createId } from "@/lib/api/create-id";
 import { DubApiError } from "@/lib/api/errors";
 import { throwIfInvalidGroupIds } from "@/lib/api/groups/throw-if-invalid-group-ids";
@@ -97,6 +98,13 @@ export const POST = withWorkspace(
       groupIds,
     });
 
+    const bountyName =
+      name ??
+      generateBountyName({
+        rewardAmount,
+        condition: performanceCondition,
+      });
+
     const bounty = await prisma.$transaction(async (tx) => {
       let workflow: Workflow | null = null;
       const bountyId = createId({ prefix: "bnty_" });
@@ -128,7 +136,7 @@ export const POST = withWorkspace(
           id: bountyId,
           programId,
           workflowId: workflow?.id,
-          name,
+          name: bountyName,
           description,
           type,
           startsAt: startsAt!, // Can remove the ! when we're on a newer TS version (currently 5.4.4)
