+++ apps/web/lib/zod/schemas/bounties.ts
@@ -28,7 +28,7 @@ export const createBountySchema = z.object({
     .string()
     .trim()
     .max(100, "Name must be less than 100 characters")
-    .nullish(),
+    .optional(),
   description: z
     .string()
     .trim()
